requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==4.11.0
django-bower==5.0.1
django-decouple==2.1
Django==1.7
gunicorn==19.1.1
ipdb==0.8
ipython==2.4.1
numpy==1.9.1
pandas==0.15.2
psycopg2==2.5.1
pycrypto==2.6.1
PyNaCl==0.2.3
pytz==2014.10
static==0.4
Unipath==1.0
wsgiref==0.1.2
